app/services/auth-service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
python-dotenv==1.0.1
psycopg[binary]==3.2.2
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1,<5
pyjwt==2.9.0
httpx==0.27.2


